requirements.txt
pytest==2.7.2
pytest-selenium
UnittestZero
